requirements.txt
Werkzeug
Flask
PyYAML